refactor PR a bit to better match existing coding style
add l10n
